requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv
sqlalchemy
psycopg2-binary
openai-agents==0.2.11
argparse